dcache-chimera: introducing parallel pool cleaning
Motivation:
Currently, pools are cleaned sequentially by cleaner. Doing so in parallel is expected to provide performance benefits.

Modification:
The number of pools that may be cleaned in parallel depends on the number of configured cleaner threads which are shared by all cleaner activities, not just this one.

Result:
Pools can be cleaned in parallel.

Target: master
Ticket: #10139
Requires-notes: yes
Requires-book: yes