srm-server: trs -- improve tape info fetching
Motivation:

When the queue is empty and a large number of requests start arriving from a bulk request, it is not ideal to directly trigger fetching tape information for the first file because the last successful fetch is sufficiently long ago.
Rather, we would like the attempt to happen regularly and the time for fetching since arrival of new requests be a bit longer on average.

Modification:

Make the tape info check happen regularly by updating the lastTapeInfoFetch on every attempt -- if there are new jobs in the queue or not.
Also move the 'sufficient time since last attempt' check into the method itself, since it it not relevant to `remove()`.
Finally, remove the condition that there need to be new jobs for attempting fetching information on tapes.

Result:

Better time distributed tape info fetching. Better handling of edge case of temporary failure to fetch tape info.

Target: master
Target: 7.2
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/13238/
Acked-by: Paul Millar