srm-server: trs fix logging
Motivation:

Active and pending tapes with jobs are logged often but are mostly empty. The same is true for the free slots which can easily be derived and do not need to be logged at all.
The logging of the queue state happens after tapes have been activated, while active and pending tapes with jobs are logged before, which might be confusing.

Modification:
Result:

Log tapes with jobs after refilling active tapes slots and only if the maps are not empty to reduce unnecessary logging.

Target: master
Target: 7.2
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/13251/
Acked-by: Svenja Meyer
Acked-by: Paul Millar