dcache: create listener, not cancel on resubmission while pinning in progress
Motivation:
When a pin requests takes too long to complete, users might retry pinning that file.
If the pin is not yet in state PINNED when that request arrives in PinManager, the original pin request will be cancelled (set to READY_TO_UNPIN, which will be removed by the regular unpinning processor) and a new pin request created in state PINNING.
This is can cause problems:
Either the pin might never be established because the request always gets cancelled before completing.
Or a pin request might have already triggered staging from tape which dCache might not be able to cancel. If the request is cancelled everywhere else in dCache, it might lead to multiple unwanted tape recalls.

PinManager should either collapse multiple pin requests with the same request id for the same pnfsid into one, or subsequent re-requests should wait for the original pin to be established.
Because not every pin requestor expects a reply once pinning has started but rather when it finishes, it is difficult to notify subsequent re-requestors via the same future.

Modification:
When a re-request arrives that asks for a pin that is in state PINNING, PinManager no longer cancels the old pin and creates a new one.
It will instead create a task that regularly checks the state of the pin of interest and returns success to the client once the pin is established.

Result:
No pin cancellations on re-requests. Instead, re-requests will wait for the original pin request to succeed and return that pin to their own client.

Target: master
Requires-notes: yes
Requires-book: yes